+++ tests/api_v2/demo_api_v2.cpp
@@ -39,8 +39,8 @@ namespace demo
     {
 // Declare entries, in this case two numbers
 #define ARILES_ENTRIES                                                                                                 \
-    ARILES_TYPED_ENTRY(real_member, double)                                                                            \
-    ARILES_TYPED_ENTRY_(integer_member, int)
+    ARILES_TYPED_ENTRY(v, real_member, double)                                                                         \
+    ARILES_TYPED_ENTRY_(v, integer_member, int)
 //         underscore ^ indicates that the name of the entry must be
 // 'integer_member_' instead of 'integer_member', this is useful if your
 // naming convention requires trailining underscores for member variables.
@@ -76,8 +76,8 @@ namespace demo
 // Ariles class (ArilesBaseClass) and a member from a non-Ariles class
 // (NonArilesBaseClass)
 #define ARILES_ENTRIES                                                                                                 \
-    ARILES_PARENT(ArilesBaseClass)                                                                                     \
-    ARILES_ENTRY_(eigen_vector)
+    ARILES_PARENT(v, ArilesBaseClass)                                                                                  \
+    ARILES_ENTRY_(v, eigen_vector)
         //              In this case ^ Ariles should not declare the inherited
         // member, therefore we use 'ARILES_ENTRY_' instead of 'ARILES_TYPED_ENTRY_'.
 
@@ -105,7 +105,7 @@ namespace demo
     class MyContainerClass : public ariles::DefaultBase
     {
         // Some of the standard containers can be used with Ariles types.
-#define ARILES_ENTRIES ARILES_TYPED_ENTRY_(myclass_vector, std::vector<MyClass>)
+#define ARILES_ENTRIES ARILES_TYPED_ENTRY_(v, myclass_vector, std::vector<MyClass>)
 #include ARILES_INITIALIZE
     };
 }  // namespace demo