+++ tests/api_v2/fixtures/003_comparison_vector.h
@@ -16,7 +16,7 @@ namespace ariles_tests
     template <class t_Configurable>
     class ConfigurableVector : public ariles::DefaultBase
     {
-#define ARILES_ENTRIES ARILES_ENTRY_(vector)
+#define ARILES_ENTRIES ARILES_ENTRY_(v, vector)
 #include ARILES_INITIALIZE
 
     public: