+++ tests/api_v2/types/simple_flags.h
@@ -16,8 +16,8 @@ namespace ariles_tests
     class ConfigurableFlags1 : public ariles::DefaultBase
     {
 #define ARILES_ENTRIES                                                                                                 \
-    ARILES_TYPED_ENTRY_(integer, int)                                                                                  \
-    ARILES_TYPED_ENTRY_(real, double)
+    ARILES_TYPED_ENTRY_(v, integer, int)                                                                               \
+    ARILES_TYPED_ENTRY_(v, real, double)
 #include ARILES_INITIALIZE
 
 
@@ -51,8 +51,8 @@ namespace ariles_tests
     {
 #define ARILES_DEFAULT_ID "ConfigurableFlags2"
 #define ARILES_ENTRIES                                                                                                 \
-    ARILES_TYPED_ENTRY_(integer, int)                                                                                  \
-    ARILES_TYPED_ENTRY_(real, double)
+    ARILES_TYPED_ENTRY_(v, integer, int)                                                                               \
+    ARILES_TYPED_ENTRY_(v, real, double)
 #define ARILES_AUTO_DEFAULTS
 #include ARILES_INITIALIZE