+++ tests/api_v2/types/simple_verbose.h
@@ -20,8 +20,8 @@ namespace ariles_tests
     class ConfigurableVerbose : public ariles::DefaultBase
     {
 #define ARILES_ENTRIES                                                                                                 \
-    ARILES_ENTRY_(integer)                                                                                             \
-    ARILES_ENTRY_(real)
+    ARILES_ENTRY_(v, integer)                                                                                          \
+    ARILES_ENTRY_(v, real)
 #include ARILES_INITIALIZE